requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
openai==1.3.7
email-validator==2.1.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
